VehicleJit now inherits from Vehicle, thus making it easier to modify in place

+++ fastsim/parameters.py
@@ -49,7 +49,8 @@ class PhysicalPropertiesJit(PhysicalProperties):
 ### Vehicle model parameters that should be changed only by advanced users
 # Discrete power out percentages for assigning FC efficiencies -- all hardcoded ***
 fcPwrOutPerc = np.array(
-    [0, 0.005, 0.015, 0.04, 0.06, 0.10, 0.14, 0.20, 0.40, 0.60, 0.80, 1.00])
+    [0, 0.005, 0.015, 0.04, 0.06, 0.10, 0.14, 0.20, 0.40, 0.60, 0.80, 1.00], 
+    dtype=np.float64)
 
 # fc arrays and parameters
 # Efficiencies at different power out percentages by FC type -- all
